refactor(app-dashboard): minor rollback, needs more work still, but works
