+++ packages/app-dashboard/src/components/withdraw/WalletConnect/QrReader.tsx
