+++ packages/app-dashboard/src/components/withdraw/WalletConnect/WalletConnectUtil.ts
@@ -33,7 +33,6 @@ export async function createWalletKit() {
 
     const core = new Core({
       projectId: '48dbb3e862642b9a8004ab871a2ac82d',
-      logger: 'trace',
       relayUrl: 'wss://relay.walletconnect.com',
     });
     walletkit = await WalletKit.init({
@@ -67,58 +66,6 @@ export async function createWalletKit() {
   }
 }
 
-export async function updateSignClientChainId(chainId: string, address: string) {
-  console.log('chainId', chainId, address);
-  // get most recent session
-  const sessions = walletkit.getActiveSessions();
-  if (!sessions) return;
-  const namespace = chainId.split(':')[0];
-  Object.values(sessions).forEach(async (session) => {
-    await walletkit.updateSession({
-      topic: session.topic,
-      namespaces: {
-        ...session.namespaces,
-        [namespace]: {
-          ...session.namespaces[namespace],
-          chains: [
-            ...new Set(
-              [chainId].concat(Array.from(session?.namespaces?.[namespace]?.chains || [])),
-            ),
-          ],
-          accounts: [
-            ...new Set(
-              [`${chainId}:${address}`].concat(
-                Array.from(session?.namespaces?.[namespace]?.accounts || []),
-              ),
-            ),
-          ],
-        },
-      },
-    });
-    await new Promise((resolve) => setTimeout(resolve, 1000));
-
-    const chainChanged = {
-      topic: session.topic,
-      event: {
-        name: 'chainChanged',
-        data: parseInt(chainId.split(':')[1]),
-      },
-      chainId: chainId,
-    };
-
-    const accountsChanged = {
-      topic: session.topic,
-      event: {
-        name: 'accountsChanged',
-        data: [`${chainId}:${address}`],
-      },
-      chainId,
-    };
-    await walletkit.emitSessionEvent(chainChanged);
-    await walletkit.emitSessionEvent(accountsChanged);
-  });
-}
-
 /**
  * Disconnect a specific WalletConnect session by topic
  * @param topic The session topic to disconnect
@@ -142,35 +89,6 @@ export async function disconnectSession(topic: string) {
     return true;
   } catch (error) {
     console.error(`Error disconnecting session: ${topic}`, error);
-    throw error;
-  }
-}
-
-/**
- * Disconnect all active WalletConnect sessions
- * @returns Promise that resolves when all sessions are disconnected
- */
-export async function disconnectAllSessions() {
-  try {
-    if (!walletkit) {
-      throw new Error('WalletKit is not initialized');
-    }
-
-    const sessions = walletkit.getActiveSessions();
-    if (!sessions || Object.keys(sessions).length === 0) {
-      console.log('No active sessions to disconnect');
-      return true;
-    }
-
-    const disconnectPromises = Object.values(sessions).map((session) => {
-      return disconnectSession(session.topic);
-    });
-
-    await Promise.allSettled(disconnectPromises);
-    console.log('All sessions disconnected successfully');
-    return true;
-  } catch (error) {
-    console.error('Error disconnecting all sessions:', error);
     throw error;
   }
 }
